predownload once

+++ predownload.js
@@ -7,23 +7,29 @@ if (!fs.existsSync('temp')) {
 }
 
 // download whole world airport
-https.get("https://raw.githubusercontent.com/davidmegginson/ourairports-data/main/airports.csv", (response) => {
-    response.on('end', function () {
-        csv()
-            .fromFile('temp/airports.csv')
-            .then((jsonObj) => {
-                fs.writeFileSync('temp/airports.json', JSON.stringify(jsonObj, null, '\t'));
-            })
+if (!fs.existsSync('temp/airports.json')) {
+    https.get("https://raw.githubusercontent.com/davidmegginson/ourairports-data/main/airports.csv", (response) => {
+        response.on('end', function () {
+            csv()
+                .fromFile('temp/airports.csv')
+                .then((jsonObj) => {
+                    fs.writeFileSync('temp/airports.json', JSON.stringify(jsonObj, null, '\t'));
+                })
+        });
+        response.pipe(fs.createWriteStream('temp/airports.csv'));
     });
-    response.pipe(fs.createWriteStream('temp/airports.csv'));
-});
+}
 
 // download fir boundary
-https.get("https://raw.githubusercontent.com/vatsimnetwork/vatspy-data-project/master/Boundaries.geojson", (response) => {
-    response.pipe(fs.createWriteStream('temp/boundaries.json'));
-});
+if (!fs.existsSync('temp/boundaries.json')) {
+    https.get("https://raw.githubusercontent.com/vatsimnetwork/vatspy-data-project/master/Boundaries.geojson", (response) => {
+        response.pipe(fs.createWriteStream('temp/boundaries.json'));
+    });
+}
 
 // download vatsim tracon boundaries
-https.get("https://raw.githubusercontent.com/vatsimnetwork/simaware-tracon-project/main/TRACONBoundaries.geojson", (response) => {
-    response.pipe(fs.createWriteStream('temp/tracons.json'));
-});
+if (!fs.existsSync('temp/tracons.json')) {
+    https.get("https://raw.githubusercontent.com/vatsimnetwork/simaware-tracon-project/main/TRACONBoundaries.geojson", (response) => {
+        response.pipe(fs.createWriteStream('temp/tracons.json'));
+    });
+}
